IPv6 support and other fixes
IPv6 support.

Now user has ability to set network dns from the
app.

build.gradle changes:
Reduced number of dependencies added in
build.gradle.
Updated dependencies to latest version.

pro-guard
dependencies for webview's JS interface changes are
removed from progaurd file.

Activities related to Webviews are removed from
AndroidManifest.xml.

New file, rethink_v12.db is added. Now, the
database builder will create the database from the
asset folder with default values.

App-wise connection details loaded in AppInfo
screen, user now can take actions against the
connection. Like block/whitelist the IP for that
particular app.

Dns queries are persisted in network logs, the value
from ipCache(cache object) will be added in network
log table.

Initial ui changes for custom domain allow/deny list.

When app name is not set for dns proxy, earlier
"Nobody" was shown. Now, if the app is not
assigned then App Name label is removed from UI.
part of #400 issue fix.

Rethinkdns will have a preloaded list of DoH
endpoint which users can directly connect to, also
there is a add icon to add new configuration.

Users now also can edit the configuration from the
user-added list.

New native screens for Rethink blocklist filetag.
(earlier it was loaded by webview activity).

Now, IP rules support both IPv4 and IPv6. Also handles IPv4 compatibile, Ipv6 prefix 64, Teredo IP's.

Notification on local blocklist download status
(success and failure).

Database migration scripts.

New filter options for Rethink blocklist configure
screen.

UI revamp.

rethink download link download.rethinkdns is
replaced with dl.rethinkdns.

+++ app/src/main/java/com/celzero/bravedns/customdownloader/DownloadFile.kt
@@ -20,8 +20,8 @@ import android.os.Parcelable
 
 class DownloadFile() : Parcelable {
     var progress = 0
-    var currentFileSize = 0
-    var totalFileSize = 0
+    var currentFileSize = 0.0
+    var totalFileSize = 0.0
 
     override fun describeContents(): Int {
         return 0
@@ -29,14 +29,14 @@ class DownloadFile() : Parcelable {
 
     private constructor(parcel: Parcel) : this() {
         progress = parcel.readInt()
-        currentFileSize = parcel.readInt()
-        totalFileSize = parcel.readInt()
+        currentFileSize = parcel.readDouble()
+        totalFileSize = parcel.readDouble()
     }
 
     override fun writeToParcel(dest: Parcel, flags: Int) {
         dest.writeInt(progress)
-        dest.writeInt(currentFileSize)
-        dest.writeInt(totalFileSize)
+        dest.writeDouble(currentFileSize)
+        dest.writeDouble(totalFileSize)
     }
 
     companion object CREATOR : Parcelable.Creator<DownloadFile> {
